Move expression optimization and type sanity checking to ExecutionState
