+++ backends/p4tools/modules/testgen/lib/execution_state.cpp
@@ -21,6 +21,7 @@
 #include "backends/p4tools/common/lib/taint.h"
 #include "backends/p4tools/common/lib/trace_event.h"
 #include "backends/p4tools/common/lib/variables.h"
+#include "frontends/p4/optimizeExpressions.h"
 #include "ir/id.h"
 #include "ir/indexed_vector.h"
 #include "ir/irutils.h"
@@ -178,10 +179,37 @@ void ExecutionState::markVisited(const IR::Node *node) {
 
 const P4::Coverage::CoverageSet &ExecutionState::getVisited() const { return visitedNodes; }
 
+/// Compare types, considering Extracted_Varbit and bits equal if the (real/extracted) sizes are
+/// equal. This is because the packet expression can be something like 0 ++
+/// (Extracted_Varbit<N>)pkt_var. This expression is typed as bit<N>, but the optimizer removes the
+/// 0 ++ and makes it into Extracted_Varbit type.
+/// TODO: Maybe there is a better way to handle varbit that could allow us to avoid this.
+static bool typeEquivSansVarbit(const IR::Type *a, const IR::Type *b) {
+    if (a->equiv(*b)) {
+        return true;
+    }
+    const auto *abit = a->to<IR::Type_Bits>();
+    const auto *avar = a->to<IR::Extracted_Varbits>();
+    const auto *bbit = b->to<IR::Type_Bits>();
+    const auto *bvar = b->to<IR::Extracted_Varbits>();
+    return (abit && bvar && abit->width_bits() == bvar->width_bits()) ||
+           (avar && bbit && avar->width_bits() == bbit->width_bits());
+}
+
 void ExecutionState::set(const IR::StateVariable &var, const IR::Expression *value) {
+    const auto *type = value->type;
+    BUG_CHECK(type && !type->is<IR::Type_Unknown>(), "Cannot set value with unspecified type: %1%",
+              value);
     if (getProperty<bool>("inUndefinedState")) {
         // If we are in an undefined state, the variable we set is tainted.
-        value = ToolsVariables::getTaintExpression(value->type);
+        value = ToolsVariables::getTaintExpression(type);
+    } else {
+        value = P4::optimizeExpression(value);
+        BUG_CHECK(value->type && !value->type->is<IR::Type_Unknown>(),
+                  "The P4 expression optimizer stripped a type of %1% (was %2%)", value, type);
+        BUG_CHECK(typeEquivSansVarbit(type, value->type),
+                  "The P4 expression optimizer had changed type of %1% (%2% -> %3%)", value, type,
+                  value->type);
     }
     env.set(var, value);
 }
